Chairs and table, sink fix, fishbowl fix
Added regular interactiveable chairs and a table
Fixed the fishbowl knife highlight error
Sink code was removed randomly so I had to reimplement it through the
trash can's code.
Made a new group called "uninter" meaning objects that are
uninteractiveable and just for show
Changed the wallpaper a little bit.